+++ shared/src/commonMain/kotlin/com/espert/reeporteciudadano/domain/model/ReportPhoto.kt
@@ -1,0 +1,3 @@
+package com.espert.reeporteciudadano.domain.model
+
+data class ReportPhoto(val id: String, val localPath: String, val exifLocation: GeoLocation?)