+++ shared/src/iosMain/kotlin/com/espert/reeporteciudadano/platform/UuidGenerator.ios.kt
@@ -1,0 +1,5 @@
+package com.espert.reeporteciudadano.platform
+
+import platform.Foundation.NSUUID
+
+actual fun generateUuid(): String = NSUUID().UUIDString()